add: TypeScriptの部分的導入
