+++ package.json
@@ -13,16 +13,21 @@
   "scripts": {
     "start": "node -r esm ./src/server/index.js",
     "server": "node_modules/.bin/webpack-dev-server --config ./webpack.config.js",
-    "babel": "nodemon ./src/server/index.js --exec babel-node --watch src -e .js,.jsx",
+    "babel": "nodemon ./src/server/index.js --exec babel-node --extensions \".js,.jsx,.ts,.tsx\" --watch src -e .js,.jsx,.ts,.tsx",
     "build": "rm -rf dist && webpack --config ./webpack.config.js --mode development"
   },
   "devDependencies": {
     "@babel/cli": "^7.10.5",
     "@babel/core": "^7.11.4",
     "@babel/node": "^7.10.5",
+    "@babel/plugin-proposal-optional-chaining": "^7.11.0",
     "@babel/preset-env": "^7.11.0",
     "@babel/preset-react": "^7.10.4",
     "@babel/preset-typescript": "^7.10.4",
+    "@loadable/babel-plugin": "^5.13.0",
+    "@types/express": "^4.17.7",
+    "@types/node": "^14.6.0",
+    "@types/react": "^16.9.47",
     "@typescript-eslint/parser": "^3.10.1",
     "babel-loader": "^8.1.0",
     "babel-plugin-module-resolver": "^4.0.0",
