+++ src/components/atoms/DateLabel/index.tsx
@@ -1,7 +1,8 @@
-import * as React from 'react';
+import React from 'react';
 
-export default function DateLabel() {
+const DateLabel: React.FC = () => {
   const hoge: string = new Date().toDateString();
-  // @ts-expect-error ts-migrate(17004) FIXME: Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
-  return <p>{hoge}</p>
+  return <p>{hoge}::::</p>
 }
+
+export default DateLabel;
