+++ src/server/index.js
@@ -1,5 +1,5 @@
 import express from 'express';
-// @ts-expect-error ts-migrate(2307) FIXME: Cannot find module 'components/pages/TopPage' or i... Remove this comment to see the full error message
+// import TopPage from 'components/pages/TopPage';
 import TopPage from 'components/pages/TopPage';
 
 const app = express();
